Address unsigned/signed warning.

+++ modules/cudaarithm/test/test_event.cpp
@@ -113,7 +113,7 @@ CUDA_TEST_P(AsyncEvent, Timing)
 {
     const std::vector<unsigned> eventFlags = { cv::cuda::Event::CreateFlags::BLOCKING_SYNC , cv::cuda::Event::CreateFlags::BLOCKING_SYNC | Event::CreateFlags::DISABLE_TIMING };
     const std::vector<bool> shouldFail = { false, true };
-    for (int i = 0; i < eventFlags.size(); i++) {
+    for (size_t i = 0; i < eventFlags.size(); i++) {
         const auto& flags = eventFlags.at(i);
         cv::cuda::Event startEvent = cv::cuda::Event(flags);
         cv::cuda::Event stopEvent = cv::cuda::Event(flags);
